fix(cosmos): simulating `verify` just means querying it
